feat: partial refactor to use filelist-utils
add static async method fromFiles
