+++ package.json
@@ -57,6 +57,7 @@
     "rollup": "^2.78.0"
   },
   "dependencies": {
+    "filelist-utils": "^0.7.1",
     "jszip": "^3.10.1",
     "openchemlib": "^8.0.0",
     "openchemlib-utils": "^2.0.0"
